feat: add sorting of winner ranking
